add: 更多的坐标轴属性控制项

+++ src/plot/KcAxis.h
@@ -57,7 +57,7 @@ public:
 
 	struct KpTickContext : public KpPen
 	{
-		double length;
+		float length;
 
 		KpPen& operator=(const KpPen& pen) {
 			style = pen.style;
